Moved the mod functions out
Two failing tests still.
